requirements.txt
astroid==2.12.11
attrs==22.1.0
autopep8==1.7.0
cffi==1.15.1
cryptography==38.0.1
dill==0.3.5.1
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
packaging==21.3
phonenumbers==8.12.56
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pylint==2.15.4
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
typing-extensions==4.4.0
wrapt==1.14.1
